Registration now opens in the browser. Clenaed up the SpectatorActivity to remove unnecessary code from earlier.
