+++ app/src/main/AndroidManifest.xml
@@ -38,7 +38,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".WebViewActivity" >
+        <activity android:name=".SpectatorActivity" >
         </activity>
     </application>
