+++ app/src/main/java/cs5150athletetracking/com/athletetracking/SpectatorActivity.java
@@ -1,0 +1,21 @@
+package cs5150athletetracking.com.athletetracking;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+
+public class SpectatorActivity extends AppCompatActivity {
+
+    private final String SPECTATOR_URL = "http://www.reddit.com";
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_web_view);
+        WebView webView = (WebView) findViewById(R.id.webView);
+        webView.setWebViewClient(new WebViewClient());
+        webView.loadUrl(SPECTATOR_URL);
+    }
+
+}
