+++ app/src/main/res/layout/activity_web_view.xml
@@ -8,7 +8,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="cs5150athletetracking.com.athletetracking.WebViewActivity">
+    tools:context="cs5150athletetracking.com.athletetracking.SpectatorActivity">
 
     <WebView
         android:layout_width="fill_parent"
